checkpoint-functions are in their own file
